news/requirements.txt
absl-py==0.9.0
gensim==3.8.3
google-api-python-client
google-cloud-storage
networkx==2.4
node2vec==0.3.2
numpy==1.18.5
pandas==1.0.4
scikit-learn==0.23.1
scipy==1.4.1
tables==3.6.1
tensorflow-enterprise-addons
torch==1.4.0
torchvision==0.5.0
transformers==3.0.0
xgboost==1.1.1
